requirements.txt
# Core dependencies
phidata>=2.0.0
python-dotenv>=0.19.0

# YouTube download
yt-dlp>=2023.11.16

# Audio processing
ffmpeg-python>=0.2.0

# Transcription and AI
openai>=1.3.0
assemblyai>=0.17.0

# Web scraping and research
duckduckgo-search>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Data handling
pydantic>=2.0.0
sqlalchemy>=2.0.0

# Utilities
python-slugify>=8.0.1
typing-extensions>=4.5.0